Refactoring AuthenticationHeader into a Navbar
