+++ client/src/App.js
@@ -5,18 +5,15 @@ import { BrowserRouter as Router, Switch } from 'react-router-dom';
 import "./App.scss";
 
 // Components
-import AuthenticationHeader from './Components/General/AuthenticationHeader'
+import Navbar from './Components/General/Navbar'
 
 import RouterObject from './Utils/Routes/Routes'
 
 const App = (props) =>  {
   return (
     <Router>
-        {/* 
-          Header with a Login / Registration Link
-          TO-DO: Create a pre-authentication navbar.
-        */}
-        <AuthenticationHeader / >
+        
+        <Navbar / >
 
       <main>
         <Switch>
